add new tables:
or_group(name, selects)

game(name,array[or_group])

A game is a pair of source and target expressions. Each of these is a CNF* expression where or_groups are sets of selects which are ORed together, 

( s1 ⋁ s2 ⋁ ..)

and where the or_groups then are conjoined together:

( or1 ⋀ or2 ⋀ ..)

to result in a single query that can be executed against all expressions.

*Conjunctive Normal Form